added-solutions

+++ 1-easy/00027-remove-element/two-pointers.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int removeElement(vector<int>& nums, int val) {
 
-        int len = nums.size();
-        int k = 0;
+        size_t len = nums.size();
+        size_t k = 0;
 
-        for (int i{0}; i < len; ++i) {
+        for (size_t i{0}; i < len; ++i) {
             if (nums[i] != val) {
                 nums[k] = nums[i];
                 k += 1;
